staging: media: atomisp: Ignore the return value of the gpio1_ctrl call
As the existing comment in the driver indicates the sensor has only 1
pin, but some boards may have 2 gpios defined and we toggle both as we
we don't know which one is the right one. However if the ACPI resources
table defines only 1 gpio (as expected) the gpio1_ctrl call will always
fail, causing the probing of the driver to file.

This commit ignore the return value of the gpio1_ctrl call, fixing this,
and add more fixes for ov2722...